+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Features**
 - [#1956](https://github.com/stripe/stripe-react-native/pull/1956) Added support for the Billie payment method
+- Added `image` to `EmbeddedPaymentElement.PaymentOptionDisplayData`
 
 **Changes**
 - Klarna: Made `billingDetails.email` and `billingDetails.address.country` optional when creating Klarna payment methods. See Stripe docs: [Migrate Klarna from Sources – Field mapping reference](https://docs.stripe.com/payments/klarna/migrate) ("Not required when using the Payment Element. It’s collected automatically.").
@@ -68,7 +69,7 @@
 ## 0.47.0 - 2025-05-21
 
 **Features**
-- Added `customerEphemeralKeySecret` and `customerSessionClientSecret` to EmbeddedPaymentElement 
+- Added `customerEphemeralKeySecret` and `customerSessionClientSecret` to EmbeddedPaymentElement
 
 ## 0.46.0 - 2025-05-08